refactor: update LocalFileNameServiceTest
