+++ naive-rpc-netty/src/main/java/com/joizhang/naiverpc/netty/serialize/SerializerType.java
@@ -2,10 +2,16 @@ package com.joizhang.naiverpc.netty.serialize;
 
 public interface SerializerType {
 
-    byte META_DATA_SERIALIZATION_ID = 1;
-    byte NATIVE_JAVA_SERIALIZATION_ID = 2;
-    byte JAVA_SERIALIZATION_ID = 3;
-    byte HESSIAN2_SERIALIZATION_ID = 4;
-    byte JSON_SERIALIZATION_ID = 5;
+    byte NATIVE_JAVA_SERIALIZATION_ID = 1;
+
+    byte JAVA_SERIALIZATION_ID = 2;
+
+    byte META_DATA_SERIALIZATION_ID = 3;
+
+    byte GSON_SERIALIZATION_ID = 4;
+
+    byte HESSIAN2_SERIALIZATION_ID = 5;
+
+    byte KRYO_SERIALIZATION_ID = 6;
 
 }
